Carve `TestDecoratedMethods` out of the monolith (wala/ML#635).
The 48 decorated/static/classmethod/same-named-method tests move
verbatim; the monolith keeps 830, preserving the 1,040 total.
